+++ fjs/asn.1/proof.f.mjs
@@ -1,5 +1,5 @@
-import { empty, isVec, length, msb, uint, unpack, vec, vec8 } from '../types/bit_vec/module.f.mjs'
 /** @import { Vec } from '../types/bit_vec/types.ts' */
+import { empty, isVec, length, msb, uint, unpack, vec, vec8 } from '../types/bit_vec/module.f.mjs'
 
 import { asBase } from '../types/nominal/module.f.mjs'