+++ fjs/basen/base128/proof.f.mjs
@@ -1,8 +1,7 @@
-import { empty, vec, vec8 } from '../../types/bit_vec/module.f.mjs'
 /** @import { Vec } from '../../types/bit_vec/types.ts' */
 
+import { empty, vec, vec8 } from '../../types/bit_vec/module.f.mjs'
 import { asBase } from '../../types/nominal/module.f.mjs'
-
 import { encode, decode } from './module.f.mjs'
 
 //