+++ fjs/bnf/descent/proof.f.ts
@@ -1,12 +1,14 @@
 import type { CodePoint } from '../../text/utf16/types.ts'
+import type { DescentMatch, CodePointMeta, DescentMatchResult } from './types.ts'
+
 import { stringToCodePointList } from '../../text/utf16/module.f.mjs'
 import { map, toArray } from '../../types/list/module.f.mjs'
 import { commaJoin0Plus, option, range, repeat0Plus, set } from '../module.f.mjs'
-import { deterministic } from '../testlib.f.ts'
 import { emptyTagMap, toData } from '../data/module.f.mjs'
 import { descentParser } from './module.f.mjs'
-import type { DescentMatch, CodePointMeta, DescentMatchResult } from './types.ts'
 import { assertEq, assertNotNullish } from '../../asserts/module.f.mjs'
+
+import { deterministic } from '../testlib.f.ts'
 
 const mapCodePoint = (cp: CodePoint): CodePointMeta<unknown> => [cp, undefined]